Shubham/Audio_emotion/requirements.txt
python-dotenv
streamlit
streamlit-audiorecorder
altair
transformers
accelerate
torch
langchain
google-generativeai
langchain-google-genai
